requirements.txt
tensorflow
torch
torchvision
opencv-python==4.2.0.34
numpy
Pillow
tqdm
easydict
tensorboardX
faiss-cpu
pandas
onnxruntime